+++ packages/soda-antd/CHANGELOG.md
@@ -1,5 +1,14 @@
 # soda-antd
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.32.0
+    - soda-hooks@6.10.1
+    - soda-react@0.0.2
+
 ## 0.7.1
 
 ### Patch Changes